# api/src/services/swagger_output.json
{
  "openapi": "3.1.1",
  "info": {
    "version": "1.0.0",
    "title": "Bingo Player API",
    "description": "Documentation of the Bingo Player API"
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
  "paths": {
...
    "/api/v1/spaces/update/by-name/{name}": {
      "post": {
        "tags": [
          "Spaces",
          "Update"
        ],
        "description": "Update an existing space in the database.",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The name of the space to update."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSpaceByName200ResponseBody"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSpaceByName200ResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              }
            }
          },
          "404": {
            "description": "No matching spaces found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpaceByNameRequestBody"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpaceByNameRequestBody"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "SpaceID": {
        "type": "string",
        "description": "The ID of a space.",
        "example": "abc123"
      },
      "SpaceName": {
        "type": "string",
        "minLength": 1,
        "description": "The name of a space.",
        "example": "My Space"
      },
      "SpaceDescription": {
        "type": [
          "string",
          "null"
        ],
        "description": "A description of a space.",
        "example": "This is a space for a Bingo Board."
      },
      "SpaceExample": {
        "type": "string",
        "minLength": 1,
        "description": "An example of what could be considered applicable to a space.",
        "example": "Example 1"
      },
      "SpaceTag": {
        "type": "string",
        "pattern": "^(in\\:)?[a-zA-Z0-9_-]+$",
        "description": "A tag for a space.",
        "example": "tag-1 or in:collection-1"
      },
...
      "CleanMongoSpace": {
        "type": "object",
        "properties": {
          "name": {
            "$ref": "#/components/schemas/SpaceName"
          },
          "description": {
            "$ref": "#/components/schemas/SpaceDescription"
          },
          "examples": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpaceExample"
            }
          },
          "aliases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpaceName"
            },
            "description": "Aliases for the space.",
            "example": [
              "Alias 1",
              "Alias 2"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpaceTag"
            }
          },
          "_id": {
            "$ref": "#/components/schemas/SpaceID"
          }
        },
        "required": [
          "name",
          "description",
          "examples",
          "aliases",
          "tags",
          "_id"
        ],
        "description": "The MongoDB representation of a Space."
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "UNKNOWN",
          "BAD_INPUT",
          "BAD_URL",
          "BAD_QUERY",
          "BAD_BODY",
          "INTERNAL",
          "NOT_FOUND",
          "NOT_IMPLEMENTED",
          "VALIDATION"
        ],
        "description": "An error code that may be returned by the server."
      },
      "ErrorMessage": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "type": "null"
          }
        ],
        "description": "A message describing an error.",
        "example": "Internal server error."
      },
      "HelpLink": {
        "type": "string",
        "format": "uri",
        "description": "A URL to the docs page that may help the user resolve the error or understand the response.",
        "example": "https://example.com/docs"
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ],
            "description": "Whether the operation was successful.",
            "example": false
          },
          "help": {
            "$ref": "#/components/schemas/HelpLink"
          },
          "code": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "description": "The error code.",
                "example": "UNKNOWN"
              }
            ]
          },
          "message": {
            "$ref": "#/components/schemas/ErrorMessage"
          }
        },
        "required": [
          "ok",
          "code",
          "message"
        ]
      },
      "ErrorResponse501": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ],
            "description": "Whether the operation was successful.",
            "example": false
          },
          "help": {
            "$ref": "#/components/schemas/HelpLink"
          },
          "code": {
            "type": "string",
            "enum": [
              "NOT_IMPLEMENTED"
            ],
            "description": "The error code."
          },
          "message": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorMessage"
              },
              {
                "example": "This feature is not yet implemented."
              }
            ]
          }
        },
        "required": [
          "ok",
          "code",
          "message"
        ]
      },
      "ErrorResponse400": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ],
            "description": "Whether the operation was successful.",
            "example": false
          },
          "help": {
            "$ref": "#/components/schemas/HelpLink"
          },
          "code": {
            "type": "string",
            "enum": [
              "BAD_INPUT",
              "BAD_BODY",
              "BAD_QUERY",
              "BAD_URL"
            ],
            "description": "The error code."
          },
          "message": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorMessage"
              },
              {
                "example": "Invalid input."
              }
            ]
          }
        },
        "required": [
          "ok",
          "code",
          "message"
        ]
      },
      "ErrorResponse404": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ],
            "description": "Whether the operation was successful.",
            "example": false
          },
          "help": {
            "$ref": "#/components/schemas/HelpLink"
          },
          "code": {
            "type": "string",
            "enum": [
              "NOT_FOUND"
            ],
            "description": "The error code."
          },
          "message": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorMessage"
              },
              {
                "example": "No resources found."
              }
            ]
          }
        },
        "required": [
          "ok",
          "code",
          "message"
        ]
      },
      "SuccessResponseBase": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "Whether the operation was successful.",
            "example": true
          },
          "help": {
            "$ref": "#/components/schemas/HelpLink"
          }
        },
        "required": [
          "ok"
        ],
        "description": "A successful response from the server."
      },
...
      "UpdateSpaceByNameRequestBody": {
        "type": "object",
        "properties": {
          "difference": {
            "type": "object",
            "properties": {
              "$set": {
                "type": "object",
                "additionalProperties": {},
                "description": "The fields to set. Each key is a path to a field.",
                "example": {
                  "name": "New Name",
                  "tags": [
                    "new-tag"
                  ]
                }
              },
              "$unset": {
                "type": "object",
                "additionalProperties": {
                  "type": "string",
                  "enum": [
                    ""
                  ]
                },
                "description": "The fields to unset. Each key is a path to a field. Cannot be a direct path, must be nested from a list field.",
                "example": {
                  "aliases.0": ""
                }
              },
              "$push": {
                "type": "object",
                "additionalProperties": {
                  "type": "object",
                  "properties": {
                    "$each": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": [
                    "$each"
                  ]
                },
                "description": "The values to push to a list field. Each key is a path to a field. Must be a direct path to a list field.",
                "example": {
                  "tags": {
                    "$each": [
                      "new-tag"
                    ]
                  }
                }
              },
              "$pullAll": {
                "type": "object",
                "additionalProperties": {
                  "type": "object",
                  "properties": {
                    "$in": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": [
                    "$in"
                  ]
                },
                "description": "The values to pull from a list field. Each key is a path to a field. Must be a direct path to a list field.",
                "example": {
                  "tags": {
                    "$in": [
                      "old-tag"
                    ]
                  }
                }
              }
            }
          }
        },
        "required": [
          "difference"
        ],
        "description": "The request body for updating a space"
      },
      "UpdateSpaceByName200ResponseBody": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "space": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CleanMongoSpace"
                  },
                  {
                    "description": "The updated space."
                  }
                ]
              }
            },
            "required": [
              "space"
            ]
          },
          {
            "$ref": "#/components/schemas/SuccessResponseBase"
          }
        ],
        "description": "The response from updating a space."
      },
...
    }
  }
}